Have both GRAB_X and GRAB_Y to "0" if not passed into get_windowinfo()
This fixes issue #1. The function get_windowinfo() is called from
do_grabxy() and if -g <WxH> was used in the invocation this
pre-setting the GRAB_W and GRAB_H variables thereby skipping the
do_grabarea() where GRAB_X and GRAB_Y were normally set. This was
causing unset variable to be passed into get_windowinfo()