Work on audio, enable audio 2 sources and more
1. Add audio settings to config file
   + pulse or alsa (need ffmpeg built with libpulse support)
   + default sources for primary and auxiliary audio sources

2. Add -a, -A and -F command line options for primary audio source,
auxiliary audio source and audio "format" (pulse or alsa).

3. Use ffmpeg's default frame rate for screencast instead if 30fps
(although art least on my system that IS 30fps.)  when it is not set
on the command line.

4. Remove '-qp 0' from the screencast ffmpeg invocation, just let ffmpeg do
what it wants. The files are a lot smaller and I can't (hardly) tell
the difference quality wise.

5. Set screencast to .mkv again

6. Maybe more that I can't remember ;-)

Also this was really only testing for screencasting so there maybe
issues with the other modes still. (Or the screencasting mode for that
matter!)